requirements.txt
Yuki Yuki Tracker application is a open source datathon project

project dependences:
	
Flask==0.11.1
Flask-Admin==1.4.2
Flask-Bootstrap==3.3.6.0
Flask-Migrate==1.8.0
Flask-MySQL==1.3
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-Sijax==0.4.1
Flask-WTF==0.12
Flask-WhooshAlchemy==0.56
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.5
PAM==0.4.2
Pillow==2.9.0
PyJWT==1.0.0
SQLAlchemy==1.0.14
SQLAlchemy-Utils==0.32.7
Sijax==0.3.2
Telerivet-Python-Client==1.0
Twisted-Core==15.2.1
Twisted-Web==15.2.1
WTForms==2.1
WTForms-Alchemy==0.16.1
WTForms-Components==0.10.0
Werkzeug==0.11.10
Whoosh==2.7.4
adium-theme-ubuntu==0.3.4
alembic==0.8.6
apt-xapian-index==0.46
argparse==1.2.1
beautifulsoup4==4.3.2
blinker==1.4
cffi==1.1.2
characteristic==14.3.0
chardet==2.3.0
click==6.6
colorama==0.3.3
cryptography==1.0.1
debtags==2.0
decorator==4.0.10
defer==1.0.6
dirspec==13.10
dominate==2.2.0
duplicity==0.7.2
ecdsa==0.13
enum34==1.0.4
future==0.15.2
html5lib==0.999
httplib2==0.9
idna==2.0
infinity==1.4
intervals==0.7.1
ipaddress==1.0.14
itsdangerous==0.24
lockfile==0.10.2
lxml==3.4.4
mysql-connector-python==2.0.4
mysql-utilities==1.6.1
ndg-httpsclient==0.4.0
oauthlib==1.0.0
oneconf==0.3.8
paramiko==1.15.2
pexpect==3.2
piston-mini-client==0.7.5
ply==3.7
pyOpenSSL==0.15.1
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.16.2
pyinotify==0.9.5
pyserial==2.7
pysqlite==2.7.0
python-apt==1.0.1
python-debian==0.1.27
python-editor==1.0.1
pyxdg==0.25
requests==2.10.0
service-identity==14.0.0
six==1.10.0
software-center-aptd-plugins==0.0.0
sqlalchemy-datatables==0.4.0
system-service==0.1.6
ubuntuone-client-data==14.4
unity-lens-photos==1.0
urllib3==1.11
validators==0.10.3
virtualenv==1.11.6
visitor==0.1.3
wheel==0.26.0
wsgiref==0.1.2
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xdiagnose==3.8.1
zope.interface==4.1.2


	
